// src/types/interface/Users.tsx
export interface Users {
  id: any;
  name: string;
  username: string;
  role: string;
}

export interface User {
  id: any;
  name: string;
  username: string;
  email: string;
  role: string;
  phone: string;
  address: string;
  city: string;
  state: string;
  country: string;
  postal_code: string;
  profile_picture: string;
}